Add coordinator for UIKit
We created a SwiftUI view that conforms to UIViewControllerRepresentable.
We gave it a makeUIViewController() method that created some sort of UIViewController, which in our example was a PHPickerViewController.
We added a nested Coordinator class to act as a bridge between the UIKit view controller and our SwiftUI view.
We gave that coordinator a didFinishPicking method, which will be triggered by iOS when an image was selected.
Finally, we gave our ImagePicker an @Binding property so that it can send changes back to a parent view.